Refactor: Update NotesViewModel and UI for improved functionality
This commit introduces several changes to enhance the NotesViewModel and the user interface:

- **NotesViewModel:**
    - Modified `loadNotes` to clear existing notes before loading new ones.
    - Updated `addNote` to correctly generate a new note ID and use milliseconds for `createdAt`. It now also calls the `addNoteUseCase`.
    - Ensured `deleteNote` only calls `deleteNoteUseCase` if the note was successfully removed from the local list.
    - Changed `notes` property to return a `toList()` copy to prevent external modification.
    - Formatted `filteredNotes` logic for better readability.
- **App.kt:**
    - Removed unused imports and simplified the main App composable by removing the initial sample note creation.
- **NotesListScreen.kt:**
    - Adjusted `LazyColumn` content padding to add more bottom space.

+++ core/src/commonMain/kotlin/me/meenagopal24/notexpert/models/Note.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Note(
-    val id: Long = 0, val title: String, val body: String, val createdAt: Int,
+    val id: Long = 0, val title: String, val body: String, val createdAt: Long,
 )
